Added ability for room owners to set a password to their rooms
Owners of a room can now do /setpassword to set a password to their room. Users will need to do /join [room] [password] to join a passworded room. Passwords are stored in a dictionary. For a focus on privacy, nothing is saved on a server restart, therefore, rooms will need to be re-created and re-passworded on server restart.